+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix: Add universal method for cleaning up LLM respones
 - Change: Switch default model to gpt4.1-mini because its much better suited
 - Feature: Add +ai tag for evaluation of tasks that can be done via LLM
+- Fix: Spot: not remembering mood and context
 
 ## [0.2.7] - 2025-08-12